Added some physic

+++ src/hs/sceneobject/Polygon.java
@@ -16,7 +16,7 @@ public class Polygon extends SceneObject
         super(gravity, origin);
     }
 
-    public Polygon(Vector2 gravity, Point2D.Float origin, int xScale, int yScale)
+    public Polygon(Vector2 gravity, Point2D.Float origin, float xScale, float yScale)
     {
         super(gravity, origin, xScale, yScale);
     }
@@ -26,9 +26,10 @@ public class Polygon extends SceneObject
         super(gravity, origin, xScale, yScale);
     }
 
-    public Polygon(Vector2 gravity, Point2D.Float origin, int xScale, int yScale, boolean hasPhysics, int numSides)
+    public Polygon(Vector2 gravity, Point2D.Float origin, float xScale, float yScale, boolean hasPhysics, int numSides)
     {
         super(gravity, origin, xScale, yScale, hasPhysics);
         this.numSides = numSides;
     }
+
 }
